+++ evalml/pipelines/components/transformers/imputers/simple_imputer.py
@@ -1,14 +1,14 @@
 """Component that imputes missing data according to a specified imputation strategy."""
 import pandas as pd
+import woodwork
 from sklearn.impute import SimpleImputer as SkImputer
-from woodwork.logical_types import NaturalLanguage
 
 from evalml.pipelines.components.transformers import Transformer
 from evalml.utils import infer_feature_types
 
 
 class SimpleImputer(Transformer):
-    """Imputes missing data according to a specified imputation strategy.
+    """Imputes missing data according to a specified imputation strategy.  Natural language columns are ignored.
 
     Args:
         impute_strategy (string): Impute strategy to use. Valid values include "mean", "median", "most_frequent", "constant" for
@@ -35,6 +35,26 @@ class SimpleImputer(Transformer):
             parameters=parameters, component_obj=imputer, random_seed=random_seed
         )
 
+    def _drop_natural_language_columns(self, X):
+        natural_language_columns = list(
+            X.ww.select(["NaturalLanguage"], return_schema=True).columns.keys()
+        )
+        if natural_language_columns:
+            X = X.ww.copy()
+            X = X.ww.drop(columns=natural_language_columns)
+        return X, natural_language_columns
+
+    def _set_boolean_columns_to_categorical(self, X):
+        boolean_columns = list(
+            X.ww.select(
+                ["Boolean", "BooleanNullable"], return_schema=True
+            ).columns.keys()
+        )
+        if boolean_columns:
+            X = X.ww.copy()
+            X.ww.set_types({col: "Categorical" for col in boolean_columns})
+        return X
+
     def fit(self, X, y=None):
         """Fits imputer to data. 'None' values are converted to np.nan before imputation and are treated as the same.
 
@@ -46,22 +66,16 @@ class SimpleImputer(Transformer):
             self
         """
         X = infer_feature_types(X)
+
         nan_ratio = X.ww.describe().loc["nan_count"] / X.shape[0]
         self._all_null_cols = nan_ratio[nan_ratio == 1].index.tolist()
 
-        # Not using select because we just need column names, not a new dataframe
-        natural_language_columns = [
-            col
-            for col, ltype in X.ww.logical_types.items()
-            if type(ltype) == NaturalLanguage
-        ]
-        if natural_language_columns:
-            X = X.ww.copy()
-            X.ww.set_types({col: "Categorical" for col in natural_language_columns})
+        X, _ = self._drop_natural_language_columns(X)
+        X = self._set_boolean_columns_to_categorical(X)
 
-        # Convert all bool dtypes to category for fitting
-        if (X.dtypes == bool).all():
-            X = X.astype("category")
+        # If the Dataframe only had natural language columns, do nothing.
+        if X.shape[1] == 0:
+            return self
 
         self._component_obj.fit(X, y)
         return self
@@ -86,20 +100,24 @@ class SimpleImputer(Transformer):
         not_all_null_cols = [col for col in X.columns if col not in self._all_null_cols]
         original_index = X.index
 
-        # Not using select because we just need column names, not a new dataframe
-        X.ww.set_types(
-            {
-                col: "Categorical"
-                for col, ltype in X.ww.logical_types.items()
-                if isinstance(ltype, NaturalLanguage)
-            }
-        )
+        # Drop natural language columns and transform the other columns
+        X_t, natural_language_cols = self._drop_natural_language_columns(X)
+        if X_t.shape[-1] == 0:
+            return X
+        not_all_null_or_natural_language_cols = [
+            col for col in not_all_null_cols if col not in natural_language_cols
+        ]
 
-        X_t = self._component_obj.transform(X)
-        X_t = pd.DataFrame(X_t, columns=not_all_null_cols)
-        if not_all_null_cols:
-            X_t.index = original_index
+        X_t = self._component_obj.transform(X_t)
+        X_t = pd.DataFrame(X_t, columns=not_all_null_or_natural_language_cols)
         X_t.ww.init(schema=original_schema.get_subset_schema(X_t.columns))
+
+        # Add back in natural language columns, unchanged
+        if len(natural_language_cols) > 0:
+            X_t = woodwork.concat_columns([X_t, X[natural_language_cols]])
+
+        if not_all_null_or_natural_language_cols:
+            X_t.index = original_index
 
         return X_t