+++ docs/source/release_notes.rst
@@ -5,6 +5,7 @@
     * Enhancements
     * Fixes
     * Changes
+        * Removed ``python_versrion<3.9`` environment marker from sktime dependency :pr:`3332`
     * Documentation Changes
     * Testing Changes
 
@@ -22,7 +23,6 @@
         * Added an ``is_cv`` property to the datasplitters used :pr:`3297`
         * Changed SimpleImputer to ignore Natural Language columns :pr:`3324`
         * Added drop NaN component to some time series pipelines :pr:`3310`
-        * Removed ``python_versrion<3.9`` environment marker from sktime dependency :pr:`3332`
     * Documentation Changes
         * Update README.md with Alteryx link (:pr:`3319`)
         * Added formatting to the AutoML user guide to shorten result outputs :pr:`3328`